requirements.txt
numpy>=1.26
pyzmq>=25.0
